Updates
Updated the code so that it always copies the users files to the disk image, and then decides whether or not to delete them. I did this so that the user would not be able to tell if the data had saved or not based on how long the process took.

Made it open a Finder window where the box is saved upon completion.

Fixed App Icon (I hope)